+++ src/context/context.service.ts
@@ -1,12 +1,8 @@
 import { Injectable } from "@nestjs/common";
-import { Context as ApolloContext } from "apollo-server-core";
-import { JwtPayload } from "jsonwebtoken";
 import { AuthTokens } from "../auth/auth.service";
 import { Claims, decodeToken, getSub } from "../auth/auth.utils";
 import { RefreshTokensService } from "../auth/refresh-tokens/refresh-tokens.service";
-import { ShieldService } from "../shield/shield.service";
 import { DataloaderService } from "../dataloader/dataloader.service";
-import { Dataloaders } from "../dataloader/dataloader.types";
 import { EventsService } from "../events/events.service";
 import { GroupMemberRequestsService } from "../groups/group-member-requests/group-member-requests.service";
 import { GroupRolesService } from "../groups/group-roles/group-roles.service";
@@ -16,35 +12,9 @@ import { PostsService } from "../posts/posts.service";
 import { ProposalActionRolesService } from "../proposals/proposal-actions/proposal-action-roles/proposal-action-roles.service";
 import { ProposalActionsService } from "../proposals/proposal-actions/proposal-actions.service";
 import { ProposalsService } from "../proposals/proposals.service";
-import { User } from "../users/models/user.model";
-import { UserPermissions } from "../users/user.types";
+import { ShieldService } from "../shield/shield.service";
 import { UsersService } from "../users/users.service";
-
-export interface ContextServices {
-  eventsService: EventsService;
-  groupMemberRequestsService: GroupMemberRequestsService;
-  groupRolesService: GroupRolesService;
-  groupsService: GroupsService;
-  imagesService: ImagesService;
-  postsService: PostsService;
-  proposalActionRolesService: ProposalActionRolesService;
-  proposalActionsService: ProposalActionsService;
-  proposalsService: ProposalsService;
-  refreshTokensService: RefreshTokensService;
-  shieldService: ShieldService;
-  usersService: UsersService;
-}
-
-export interface Context extends ApolloContext {
-  claims: {
-    accessTokenClaims: JwtPayload | null;
-    refreshTokenClaims: JwtPayload | null;
-  };
-  loaders: Dataloaders;
-  permissions: UserPermissions | null;
-  services: ContextServices;
-  user: User | null;
-}
+import { Context, ContextServices } from "./context.types";
 
 interface RequestWithCookies extends Request {
   cookies?: { auth?: AuthTokens };
