+++ src/context/context.types.ts
@@ -1,0 +1,43 @@
+import { Context as ApolloContext } from "apollo-server-core";
+import { JwtPayload } from "jsonwebtoken";
+import { RefreshTokensService } from "../auth/refresh-tokens/refresh-tokens.service";
+import { Dataloaders } from "../dataloader/dataloader.types";
+import { EventsService } from "../events/events.service";
+import { GroupMemberRequestsService } from "../groups/group-member-requests/group-member-requests.service";
+import { GroupRolesService } from "../groups/group-roles/group-roles.service";
+import { GroupsService } from "../groups/groups.service";
+import { ImagesService } from "../images/images.service";
+import { PostsService } from "../posts/posts.service";
+import { ProposalActionRolesService } from "../proposals/proposal-actions/proposal-action-roles/proposal-action-roles.service";
+import { ProposalActionsService } from "../proposals/proposal-actions/proposal-actions.service";
+import { ProposalsService } from "../proposals/proposals.service";
+import { ShieldService } from "../shield/shield.service";
+import { User } from "../users/models/user.model";
+import { UserPermissions } from "../users/user.types";
+import { UsersService } from "../users/users.service";
+
+export interface ContextServices {
+  eventsService: EventsService;
+  groupMemberRequestsService: GroupMemberRequestsService;
+  groupRolesService: GroupRolesService;
+  groupsService: GroupsService;
+  imagesService: ImagesService;
+  postsService: PostsService;
+  proposalActionRolesService: ProposalActionRolesService;
+  proposalActionsService: ProposalActionsService;
+  proposalsService: ProposalsService;
+  refreshTokensService: RefreshTokensService;
+  shieldService: ShieldService;
+  usersService: UsersService;
+}
+
+export interface Context extends ApolloContext {
+  claims: {
+    accessTokenClaims: JwtPayload | null;
+    refreshTokenClaims: JwtPayload | null;
+  };
+  loaders: Dataloaders;
+  permissions: UserPermissions | null;
+  services: ContextServices;
+  user: User | null;
+}
